Models no longer use UUID as primary_key
Issue: UUID could not be changed without creating a duplicate plant/tray entry. When the primary_key is changed it no longer matches the original row, so a new row is created when saved. This made it impossible to assign an existing plant/tray a new QR code.

Fix required replacing old migrations, no viable approach to migrate existing database. When id is removed django automatically generates a migration replacing it with BigAutoField, but this cannot accept UUID (not valid base64). Tried manually writing migrations to copy uuid to a new UUID field (works), generating sequential integers in temporary IntegerField (works), and then copying the integers to a BigAutoField with primary_key=True, but this broke all ForeignKeys. Could have kept primary_key as a uuid in addition to the uuid attribute (confusing) to preserve database, but fortunately caught before anything but test data was added.