Fix failing tests
Root issue of all failures (except snapshot): setTimeout lines are not called on my dev machines, but are called in the gitlab-runner environment. This is why lack of scrollIntoView mock didn't make tests fail on dev machine (was never called). The lines inside setTimeout in toast effect also aren't called on dev machines, but are in the pipeline, apparently after the component has unmounted and the ref is null.