Bug fix: Photos with no DateTimeOriginal exif param no longer break
Issue: When DateTimeOriginal is not set the created field was populated by copying the uploaded field in the save method. However, the uploaded field is not set until the end of the save method (by auto_now_add), resulting in created being set to None. This broke the __str__ method and caused an uncaught exception when manage_plant page was loaded.