Fix test failing due to inadequate system time mock
Issue: luxon.DateTime.now was globally mocked to return a predictable value, but the builtin Date was not mocked (except with fakeTimers in snapshots). The react-calendar component uses Date to determine the visible month, but most of the other components use luxon. This caused a manage_plant regression test ('updates calendar when repot modal is submitted') to fail because an event marker (uses luxon mock == March) was added to a non-visible calendar page (April visible due to lack of Date mock).

Solution: Date is now globally mocked to the same value as luxon.DateTime.now, fakeTimers are no longer required for snapshot tests