NoteModal now supports editing existing notes (API not implemented)
Not happy with this approach but seems to have fewest tradeoffs. The useEffect causes previous contents to be visible for 1 frame, but without it old contents persist when editing a different note. This can be fixed by moving textarea to separate component, but then re-render makes textarea lose focus while typing. There was another solution that I can't remember which caused buttons to animate on every keystroke. Should refactor later.