Bug fix: manage_plant no longer shows outdated group name if group was renamed since cache saved
Issue: Saving group does not update cached state for each plant in group. This could be fixed with a hook for named groups, but unnamed groups would still be outdated (index changes if other groups renamed). It might be worth rethinking the sequential unnamed plant/group names, several things could be simplified if these weren't shown on the frontend.