Prevent creating events for archived plants using FilterColumn radio buttons
Issue: Water/Fertilize all buttons skipped archived plants since 3876113a, but it was still possible to manually select an archived plant using the FilterColumn radio buttons and create an event. These events could not be deleted through the UI (no delete events option on manage_plant page if plant is archived).

UX could be better - ideally the radio button should disappear or be disabled for archived plants (instead of just ignoring the user's selection), but this can't be done without a major refactor