Bug fix: no longer creates duplicate cards if AddPlantsModal used twice
Issue: The selected refs in AddPlantsModal and RemovePlantsModal were not cleared in the submit handler. If plant1 was added, then the add modal was opened again and plant2 was added, a second plant1 card would appear since the UUID was still in the selected ref. This was impossible to fix without reloading the page since the checkbox to add the UUID was unrendered after submit. RemovePlantModal issue was not visible to user but still posted UUID of plant that was already removed if submitted twice.

Issue present since before react refactor, unnoticed because using add/remove twice in the same session is very rare.