Fix modals filling full height on iOS safari
Same style as vanilla daisyui but with dvh instead of vh. On iOS safari vh is equal to heght when address bar is hidden, so when address bar is visible the modal hits top and bottom of the screen. With dvh (dynamic) modal has the same y margin whether address bar is visible or not.