Bug fix: Plants removed from group no longer stay selected in FilterColumn
Issue: FilterColumn adds selected plant UUIDs to selectedPlants ref and only removes them when the water/fertilize buttons are clicked. If a plant was selected and then removed from group before the event buttons were clicked it's UUID would still be in selectedPlants. This caused an exception in removeArchivedPlants (plantDetails.find) which prevented creating events until the page was refreshed.