Refactor EditableNodeList to use form to read selection
No performance impact, reduces chance of bugs like 617e0ec3 where selectedRef contents become outdated. Instead of updating a ref each time a node is selected the whole array is read once on submit with FormData. No longer have to clear selection when a node is deleted (once it unrenders it's checkbox won't be in the form).