Increase clickable area to full node width
e6231192 left a small area on the right side that couldn't be clicked to avoid blocking PlantCard/GroupCard details buttons (ie card could be selected by clicking anywhere to left of button, but not to the right, above, or below). Now that clip-path was removed in aed44879 (caused some kind of stacking context issue) the z-index on PlantCard/GroupCard button positions it above EditableNodeList label, so the button can be clicked even with the label set to full width.