Bug fix: No longer fails to add first event on day with existing notes/photos
Issue: timelineSlice.eventAdded checked if dateKey existed in timelineDays and assumed it would also exist in calendarDays - but it is possible for timelineDays to have extra keys if they only contain photos and/or notes. This resulted in an error when attempting to write to .events attribute of a nonexisting key in calendarDays, which prevented either state from being updated.