Bug fix: Deleting group no longer deletes plants in group
Issue: Plant.group (ForeignKey) had on_delete set to CASCADE, causing all plants in a group to be deleted when the group was deleted. There was no warning for this when deleting a group in the frontend. Can't believe this went unnoticed for so long (don't think I've ever deleted a group in prod).