Sort "Never watered" as least-recent, not most-recent
Standard behavior is to put null values at the end but this doesn't make sense when sort key is watered (puts never watered next to most-recent). When users sort by last watered to see which plants they need to water plants which have never been watered should be at the top of the list.