Move defaultPhoto to timelineSlice
Keeping it in plantSlice initially seems simpler (it's already in plantDetails) but gets very complicated when plant default photo is not set (will default to most-recent photo, but photo uploads are handled in timelineSlice which can't dispatch an event to plantSlice). With the old setup changing the default photo when photos were uploaded was going to require a default_photo_set bool in the backend state and corresponding frontend stuff, but with all photos in timelineSlice it's fairly simple.