Use bulkCreateMockContext in tests that don't require all keys
If only the required keys are included but a new key is added later then dozens of test files need to be edited and the diff becomes hard to read. Use approach from 42a5d90f to ensure entire mockContext is created in all tests where possible.

Improve targetting in EventHistoryModal tests.