Bug fix: Clicking GroupModal options no longer redirects to manage_group page
Broken since ae9d90fc removed the linkPage param (reintroduced manage page links to the cards in GroupModal). In all other cases the card was covered by a transparent div which would prevent the card from being clicked, but in GroupModal the card was a child of a div with click listener. When clicked the /add_plant_to_group call was made then immediately canceled when the click propagated down to the manage link.