Block non-numeric chars in pot_size field, show number keyboard (mobile)
If alphabetical characters were entered in field (type="number") FormData would contain an empty string instead of the user's value. This could easily go unnoticed and could happen from a typo.

Change field type to text (on Firefox and iOS Safari the onInput clears the whole field when a letter is pressed if the type is number, but not if it's text).