Don't add plants that are in another group to manage_group AddPlantsModal options
Selecting a plant that was already in another group worked but removed it from the existing group, which may not be what the user expected. This also made the list very long, even if most plants were already in groups.

Would be a lot easier to update options state if it was an object keyed by uuid...