Revert "Per-page tailwind tree shaking (not project-wide)"
This reverts commit e378672edb70c5a34f28ff92f6dada118d944554.

Knowing how to do this will probably be useful in the future (eg if non-public admin pages are added, or pages that are only available to certain users) but it doesn't make sense for an app like this where most users will load every single page. The bundles are very small but they're also unique and need to be cached separately, and the total filesize is 320KB. Most of that is duplication (common tailwind classes, theme vars, etc) that are included in each bundle. It's better than having larger identical bundles with different filenames (still cached separately), but there are more efficient ways to do it (if all the shared CSS is extracted to a reusable bundle the total size drops under 100KB).