Fix FilterColumn disappearing when user types in input
Issue: Since header with title and filter input is sticky it is possible for the actual top of column to be above viewport while the header (visually the top) is visible. When contents are filtered column height changes, and if the new column height fits in the offscreen part (whole column now offscreen) the header unsticks and the whole column disappears. This could be very confusing on mobile overview page.

Solution: instantly scroll top of column to same position has header when filter query changes (prevents column going offscreen, user doesn't see scroll since sticky header does not move).