Bug fix: Allow uploading RGBA photos
Issue: photo thumbnails are saved as JPEG, which is not compatible with transparency. If an RGBA photo was uploaded an unhandled exception would occur while creating the thumbnail, causing /add_plant_photos to return 500. This prevented all photos in the payload from being saved, not just the RGBA photo.

The photo model now checks image mode before attempting to create the thumbnail and converts RGBA into RGB.

Closes issue #6