Bug fix: prevent unnamed plant index changing
Issue: get_unnamed_plants (and groups) used an unsorted queryset, assuming it would be in the same order plants were added to database. This is true on sqlite but not postgres. Adding a a ForeignKey to a plant seems to reliably move it lower in the queryset than plants without the ForeignKey, causing unnamed plant index to change unexpectedly when the plant was added to a group.

Both get_unnamed_plants and get_unnamed_groups now sort by created time (match expected order). Can't find a way to make group index change for regression test but probably possible (changing user would do it but there is currently no way to do that).