Bug fix: Cached group details outdated if plant in group deleted
Issue: Cached group details only updated when the Group model was saved/deleted (hooks) or when plants were added/removed (update functions called in views). If a plant in group was deleted (number of plants in group decreases) the cached state did not update.

Solution: add conditional to plant save/delete hooks to check if plant is in group and update caches. No longer need manual update functions in add plants to group endpoints (but still need in remove, hook can't detect the group plant used to be in).