Fix note height not updating when edited
Issue: When note is expanded it's height is set to a fixed value. If the note is then edited and becomes taller the height did not update, so the new lines were truncated until the user collapsed and re-expanded the note.

Can't really regression test this (would have to mock scrollHeight, and then you're just testing mocks). Can reproduce by creating short note, expanding, then editing to much longer note.