Don't cache plant_options or group_options
These are now only served when user opens respective modals (not in initial state, so doesn't block page load) and take 5ms (plant_options) and 1ms (group_options) to build. Modals are opened infrequently but actions that caused cached options to update are common, so this is almost guaranteed to be a net loss (ie adding cumulative 20-100ms of updates just to save 5ms when modal finally opened).