Benchmark all endpoints
Not looking good for cached state updates, although most of the extra queries are due to signals looking up the models again. In shell I can write a known last_watered timestamp to cached overview state and re-cache in 0.25ms, so updating them directly in views might be worth it.