Bug fix: cached manage_plant state not deleted when uuid changed
Once UUID is changed the cached state cannot be accessed (key contains old uuid), so this is just a waste of memory (and potentially results in wrong page being shown if a different plant is registered with that uuid).

Discovered while writing tests to confirm that cached states update correctly when endpoints are called (not when models are saved), will allow removing/modifying signals without breaking cached updates.