Show event timestamp(s) on hover
timelineDays events key now contains objects wis with type and timestamp keys (instead of just type strings), contains all events if multiple of same type exist on same day. This allows passing all timestamps to EventMarker component. The calendarDays state is unchanged (but is built differently).

The real point of this is to allow selecting events to delete from the timeline, tooltips are just a bonus. Couldn't select events before because only had YYYY-DD-MM dateKey of each event (user timezone, but events to delete are in UTC - no easy or efficient way to determine which UTC timestamps match a user-timezone dateKey, have to iterate and convert all, wont scale etc).