Add instructions to overview FloatingFooter
Now tells user to select plants and groups, shows number selected, shows "Hold to confirm" in footer instead of tooltip. Instructions are less necessary here since the checkboxes make it pretty clear, but the counter is useful (if user exited edit mode and re-entered they might not know things are still selected and delete accidentally - should probably switch selection to state so it can be cleared).