Replace selected note strikethrough with outline
Consistent with other elements, better transition. Wanted to animate strikethrough from left to right but can't do that without pseudo element (no line wrap) or background image hack (works but draws line under text not on top). Probably some complicated way to do it with background if another absolute-positioned div with transparent text is on top or something but not worth it for now.