Confirm that link to scanned QR code appears
Still can't get coverage for the onError callback. This library is a nightmare to mock, to the point where I'm considering switching even though it works great otherwise. To trigger onError navigator.mediaDevices.getUserMedia needs to be mocked to raise a DOMException, but it propagates and makes the test fail, and adding expect().toThrow() or wrapping stuff in try/catch doesn't fix it.