Custom notification sounds
Play success sound when URL is part of app, error sound when unrelated URL is scanned.

Sounds are from react-sounds package, was going to use that but they are not included in dist (pulls from CDN instead). Just downloaded them and included in bundle, switched to use-sound which is smaller and more widely used.

Also no longer shows open button when an unsupported URL is scanned.