Fix django-debug-toolbar in development-docker-compose.yaml
Using SHOW_TOOLBAR_CALLBACK to circumvent INTERNAL_IPS (won't run in production, requires debug mode). Not ideal but can't get it to work in docker otherwise. Upgraded to 6.0.0 to use the new show_toolbar_with_docker function which is supposed to fix this but it doesn't work on linux (host.docker.internal doesn't exist, after adding it with extra_hosts: - "host.docker.internal:host-gateway" it does but resolves to a different IP than the one request comes from).