Fix unable to log out
Caused by indescriminately replacing <a> with <link>. Since /accounts/logout/ is not in routes.js this fell back to the default route (/) and navigated to overview without sending request to backend.

Could fix this within SPA context with redirect but probably safest to just do a full page load so any leftover state is cleared (user email etc is on profile page, might be possible to leak that if logged out without leaving SPA session and then someone else used same tab).