Fix development docker compose
This has been broken since 0a58b2ba and was masked by the persistent volumes already existing. When the celery line ending with & (background) was added at the end of a chain of statements separated by && the entire chain was backgrounded, causing the final line to run immediately (before dependencies were installed). On dev machine this wasn't a problem since dependencies were already in the volume, but if that volume was deleted (how I discovered it) it would exit immediately with an import error (django not found).

Remove && chain (unnecessary, it will still fail just with more errors, but less chance of something like this happening again). Move requirements.txt to /tmp (was writing to directory where repo is mounted, so requirements was written in repo and would be mounted in again next run).

Document how to clear persistent volumes.