Fix SPA doing full-page reload after scanning some QR codes
Issue: was creating <Link> with full URL in the to= prop, internal navigation is only guaranteed when to= is a relative path. Didn't fully debug why it happens on some QR codes but not others, guessing some of the old ones have http instead of https.

This actually would have fixed the repeated camera permission prompt without converting app to SPA... Oh well, worth it for other performance improvements.