Disable pinch zoom in QR scanner
Very unintuitive, user is obviously trying to zoom the camera but pinch zoom just zooms the page like usual (which makes all the buttons including close go off screen). Ideally pinch zoom would do the same thing as zoom in/out buttons but the library doesn't provide a way to do that, so disabling for now.