Bug fix: Still able to add events to archived group
When a plant is archived the event buttons are replace by text saying "Plant Archived", but when a group was archived nothing changed. Users were still able to log events and there was no visual indication that the group was archived. Groups now replace the buttons with "Group Archived" to match the manage plant behavior.

Closes issue #27