Add optional onTitleClick prop
Can be used to show something other than dropdown when title is clicked (dropdown is limiting, animation is generic and there is no close animation).

Make menuOptions required (doesn't actually have conditional that checks if given, if skipped there will just be a button that opens an empty dropdown div).