Keep dropdown button highlight visible while menu focused
Old selector only kept the highlight on the button while the button was focused, once focus moved to menu it disappeared. This would happen when an option was clicked in a dropdown that doesn't close automatically (ie FilterColumn sort menu, top-left corner menu when theme option is clicked). Button highlight should stay visible to indicate that the menu is open.