Fix FloatingFooter blocking bottom 1rem when closed
Issue: translates a fixed height off screen, which was correct when originally written, but height increased when text was added. This caused 1rem of FloatingFooter to cover the document (invisible because of opacity-0) and prevent elements under it from being clicked.

Really should switch to a non-fixed distance but when I tried it the animation looked worse, need to spend more time tweaking it.