Scroll timeline to top of page when quick nav opened
Ensure quick nav menu options are visible, fixes bad UX but in this case we can't scroll bottom of menu into view (possible for menu to be taller than viewport, so that would just move to an arbitrary place in timeline without showing menu which is sticky). Considered having a listener that scrolls when year options expanded but that could cause scroll to jump around a lot, simpler to just align timeline title to top (user is browsing timeline if they opened this, so showing full timeline makes sense).