Fix visual flash when last plant/group unarchived
Issue: state update causes columns to disappear and setup instructions to be rendered before redirecting to main overview. Could skip by skipping state updates and redirecting immediately when the number of selected plants/groups matches total archived plants/groups, but that breaks if BOTH plants and groups are unarchived simultaneously (the plantsRemoved/groupsRemoved handlers run in parallel and see that the other state is not empty yet, so page won't redirect). Simpler to just only render setup instructions on main overview (slight flash when columns disappear but that will be fixed when transition is added).

Closes issue #21