Fix slightly asymetrical circle around logo
Issue: circle_size is the absolute width (px) of the overlay image. When the circle_mask is drawn (draw.ellipse) circle_size was used as the right and bottom edges, but since the top and left edges are 0-indexed this overflows the image by 1px. This resulted in a "flat spot" on the right and bottom edges of the circle (1px row was cut off).