Align logo to QR code grid
Instead of arbitrarily scaling logo to 35% of QR width (will slice off part of adjacent modules on all edges) scale it to the nearest value which is divisible by qr_scale (width of each QR code module). This way the edges of the logo perfectly align to grid (if logo was a square all adjacent modules would be either fully occluded or fully visible).

Reduces chance of thin lines next to logo where the logo padding covered most of a QR code module, but not all of it (tidier appearance, less unscanable modules). This mostly happened on the right and top edges where the logo is mostly a straight line (impossible to get thin line here now, but still get diagonally sliced modules next to curved parts of logo but those don't look as bad).