Add option to only match QR codes with unregistered UUIDs
Will be used to change QR codes of existing plants/groups (scan QR code, click confirm, never have to load register page). In this scenario we only want to match unregistered QR codes, not everything with the right domain.

When a QR code is matched handleScan waits on an /is_uuid_available API call and only shows the button if it returns available. highlightQrCodes shows a grey highlight while waiting on the API call and switches to green/red once resolved. Responses are cached so that only 1 request is made per URL (don't DDOS backend with 1 request per frame).