Bug fix: Prevent changing visible page with tab key
Issue: since all pages are rendered (even if not visible) it was possible to move focus from an input on the visible page to an input on one of the non-visible pages by pressing the tab key. This shifted the non-visible page into the viewport (usually not aligned to edge of page) and could be used to bypass disabled next button.