Add default database + redis connection settings
Postgres + redis hostnames are hardcoded in docker-compose healthcheck and their own containers, image should default to the same strings. No reason to require env vars that might confuse the user (and reak everything if default value changed). If the user wants to change this for some reason they already have to understand healthcheck and update it everywhere, advanced users like that can add the env var if they really need to.