No longer need to remove duplicates (always unique)
This was necessary when events was an array of type strings ("water", "fertilize", etc) that might contain duplicates if multiple events of same type existed on that day (needed to remove to avoid duplicate EventMarker in timeline). After 616fefd2 changed it to an array of objects that contain timestamps it is no longer possible for there to be duplicates (can't have 2 events with same timestamp, enforced by database).