Fix NoteCollapse select outline (delete mode) filling full width
Now matches note text width, does not draw excessively wide outline on short notes.

The only downside to this fix is it causes inconsistent alignment when there are long and short notes next to each other. The scale down animation causes the left edge of the text to shift right (text width shrinks). When the wrapper div is full width all notes shift right the same amount (start stays aligned), but with this fix short notes shift less (looks less indented next to long note). Realistically this is very uncommon in usage, especially since notes have to be very short on mobile to not line wrap.