Add Remove QR Code button to archived plants
Replaces Change QR Code button when plant is archived, opens confirmation modal that changes to random UUID if user confirms

Use case: User archived a plant, empties pot but leaves sticker on it, later pots a different plant in that pot and tries to register it but finds QR code is already connected to the archived entry. Can now just click remove to free the UUID and register it (without having to change it to a different QR code which then can't be used either like before).