Bug fix: Footer text incorrect if selection changed right after releasing button
Issue: Since footer shows "Hold to confirm" for 750ms after release to give user who tapped a chance to see it, its possible to change selection during this window before it changes back to "n items selected". When this happens totalSelected updates immediately, but HoldToConfirm's onHoldStop callback prop contains a closure with the totalSelected when the timeout started. This caused the footer to change back to the number that were selected when click released even if it had changed.

Solution: set footer text from ref with a copy of totalSelected (updates immediately when totalSelected changes, but closures get current value unlike plain var).