Remove Plant queryset with_uuid_as_string_annotation method
This is slower than converting the uuid instances to str in a list comprehension (and shifts work from django to the database, easier to scale the other way). Not a huge difference but consistently slower whether I'm querying 4, 10, or 100 entries. On dev machine it's about 1% slower when adding overview annotation but up to 7% slower when only annotating last_watered/fertilized. On VPS I measured 9% slower with overview annotation, likely worse with smaller annotation.