Add DetailsChangedEvent model
Records plant details before and after /edit_plant_details API call. Instantiates 1 model per day in user's timezone (not UTC). If model on current day already exists it is loaded from database and updated (overwriting _after values from the first edit on that day). Goal is to only track details at the start and end of the day, not literally every change (if user changes name multiple 3 times we don't want 3 markers in the timeline, this would just punish users for making typos).

Plan is to eventually add uuid_before/after and archived_before/after fields that update when QR code is changed or plant archived/unarchived, will most likely have a decorator to query/create the DetailsChangedEvent and pass to endpoints that need it.