Bug fix: renders empty TimelineDay for DetailsChangedEvent with no changes
Issue: if user changed name/species/etc and then changed back to original value on the same day a DetailsChangedEvent will exist with matching _before and _after values. This caused a timelineDays entry to be created which rendered an empty TimelineDay containing an empty DetailsChangedSection. These are now only added to the state if changes exist, and timelineDays day is removed if it only contains a change event with no changes.